+++ src/test/java/nextstep/member/acceptance/MemberAcceptanceTest.java
@@ -1,11 +1,13 @@
 package nextstep.member.acceptance;
 
+import nextstep.member.utils.AuthSteps;
+import nextstep.member.utils.MemberSteps;
 import nextstep.utils.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import static nextstep.member.acceptance.MemberSteps.*;
+import static nextstep.member.utils.MemberSteps.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class MemberAcceptanceTest extends AcceptanceTest {
